+++ occtax/install/sql/upgrade/upgrade_2.12.8_2.12.9.sql
@@ -1,0 +1,3 @@
+-- Index sur le champ odata de occtax.observation
+DROP INDEX IF EXISTS observation_odata_import_time;
+CREATE INDEX observation_odata_import_time ON occtax.observation USING gin (odata);